Fill gaps in budget dict with derived terms
Major changes:
- The xarray dataset containing diagnostics is not passed to the xbudget
functions so that derived terms can be (lazily) added to the dataset. The names of derived
variables follow from the nested dictionary structure and are unambiguous.
- dictionary structure has been modified so that the names of the new derived variables are also
added to the dictionary. This facilitates evaluation of the reconstructed terms as it is now obvious
if there are existing diagnostics that they should be equivalent to.
- the collect_budgets function fills the budget with derived terms and the `aggregate` function
is used to collect just some high-level terms

To do:
- Functions for more flexible grouping/aggregating/decomposing of budget terms. For example, what if someone wants a
  high-level budget for most terms but wants to really break down the surface fluxes as much as possible? They currently
  can do this manually by following the budget structure, but it would be nice to have helper functions to make it easier.
- Support derived variables that are finite-differences of fluxes, to be compared against the flux-divergence diagnostics.
- Somehow protect against users combining inconsistent terms (e.g. on different grids).