Corrected sign of salt_mass_flux term in MOM6 conventions file!
This bug slipped past me because the term is 1) quite small in the
Baltic Test Case and 2) was correctly specified (manually) in the
2D surface WMT notebook that replicated @jetesdal's previous results.

I discovered the bug because in some preliminary analysis of water mass
budgets in density space south of 30S, the surface salt flux term was a
leading order error (clearly the wrong sign, producing lighter waters in
Antarctic winter) and projected into a leading order error in the
spurious mixing / cabelling term. Those calculations are now fixed.

- Re-ran the notebooks with the fixed inputs.