Implements calculations of flux convergences from fluxes on cell faces
Version bumped to v0.0.3. This change is relatively major.
It should not be a breaking change but I did not careful test whether it is.

This functionality leverages the `xgcm.Grid.diff` method that enables
grid-aware finite-differencing. We introduce the new `difference` category
of entries in the `xbudget` budget dictionary.

As an initial proof of concept, we add an entry to `MOM6.yaml` that shows
how to compute the lateral convergent mass transport into a cell (or "layer").

Since this new functionality requires an `xgcm.Grid` instance that details
the structure of the model grid, I add `xgcm` as a dependency (for now
pinning to the same version that I have pinned `xwmt` and `xwmb` to).